+++ common/src/main/java/club/iananderson/seasonhud/client/gui/screens/DisplayOptionsScreen.java
@@ -26,6 +26,7 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
   private Location hudLocation;
   private int posX;
   private int posY;
+  private int fabricSeasonsRow;
   private double seasonScale;
   private ShowDay showDay;
   private boolean seasonColor;
@@ -55,7 +56,7 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
   }
 
   public void loadConfig() {
-    drawDefaultHud = Common.drawDefaultHudMenu(this.minecraft);
+    drawDefaultHud = Common.drawDefaultHudMenu();
     hudLocation = SeasonHudClient.getHudLocation();
     posX = SeasonHudClient.getHudX();
     posY = SeasonHudClient.getHudY();
@@ -129,17 +130,19 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
     super.render(graphics, mouseX, mouseY, partialTicks);
     MutableComponent seasonCombined = CurrentSeason.getInstance(this.minecraft)
         .getConfigText(showDay, showSubSeason, seasonColor);
+    posX = Client.DEFAULT_X_OFFSET;
+    posY = Client.DEFAULT_Y_OFFSET;
 
     if (drawDefaultHud) {
+      hudScaleSlider.visible = true;
+      sliderX.visible = true;
+      sliderY.visible = true;
+
       boolean customLocation = (hudLocationButton.getValue() == Location.CUSTOM);
-      hudScaleSlider.visible = drawDefaultHud;
       sliderX.active = customLocation;
-      sliderX.visible = drawDefaultHud;
       sliderY.active = customLocation;
-      sliderY.visible = drawDefaultHud;
 
       seasonScale = hudScaleSlider.getValueDouble();
-
       int componentWidth = (int) (this.font.width(seasonCombined) * seasonScale);
       int componentHeight = (int) (this.font.lineHeight * seasonScale);
 
@@ -178,18 +181,6 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
       }
     }
 
-    if (Common.fabricSeasonsLoaded() && Common.clientSideConfig(this.minecraft)) {
-      int row = 5;
-
-      if (!drawDefaultHud) {
-        row -= 2;
-      }
-
-      GuiComponent.drawCenteredString(graphics, font, "Day Length", leftButtonX + buttonWidth / 2,
-                                      MENU_PADDING + (row * (buttonHeight + BUTTON_PADDING)) - (font.lineHeight
-                                          + BUTTON_PADDING), 16777215);
-    }
-
     if (Common.sereneSeasonsLoaded()) {
       fertilityReplacesSeasonButton.active = showFertility;
     }
@@ -198,6 +189,11 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
     graphics.translate(0, 0, 50);
     graphics.scale((float) seasonScale, (float) seasonScale, 1.0F);
     GuiComponent.drawString(graphics, font, seasonCombined, posX, posY, 0xffffff);
+
+    if (Common.fabricSeasonsLoaded() && Common.clientSideConfig(this.minecraft)) {
+      drawColumnHeading(graphics, Common.literalText("Fabric Seasons Day Length"), Side.LEFT, (fabricSeasonsRow));
+    }
+
     if (CurrentFertility.getInstance(this.minecraft).shouldDrawNewLine()) {
       MutableComponent fertility = CurrentFertility.getInstance(this.minecraft).getHudText();
 
@@ -221,17 +217,10 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
 
   // TODO: Make subscreens for Season options, fertility options. Make sure the preview displays on both
 
-  @Override
-  public void init() {
-    loadConfig();
-    super.init();
-
-    MutableComponent seasonCombined = CurrentSeason.getInstance(this.minecraft).getHudText();
-
-    row = -1;
-
+  private void defaultHudButtons() {
     if (drawDefaultHud) {
-      row += 1; // Row 1
+      row += 1;
+
       hudLocationButton = CycleButton.builder(Location::getLocationName)
           .withTooltip(t -> Common.newTooltip("menu.seasonhud.season.hudLocation.tooltip"))
           .withValues(Location.values())
@@ -247,7 +236,9 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
           .withBounds(rightButtonX, (buttonStartY + (row * offsetY)), buttonWidth, buttonHeight)
           .build();
 
-      row += 1; // Row 2
+      row += 1;
+
+      MutableComponent seasonCombined = CurrentSeason.getInstance(this.minecraft).getHudText();
       sliderX = HudOffsetSlider.builder(Common.translatedText("menu.seasonhud.season.xOffset.slider"))
           .withValues(0, this.maxWidth(seasonCombined), posX, Client.DEFAULT_X_OFFSET)
           .withBounds(rightButtonX, (buttonStartY + (row * offsetY)), buttonWidth / 2 - BasicSlider.SLIDER_PADDING,
@@ -262,8 +253,11 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
 
       widgets.addAll(Arrays.asList(hudLocationButton, hudScaleSlider, sliderX, sliderY));
     }
+  }
 
-    row += 1; // Row 3 (enableMinimapIntegration -> Row 1)
+  private void seasonButtons() {
+    row += 1;
+
     CycleButton<ShowDay> showDayButton = CycleButton.builder(ShowDay::getDayDisplayName)
         .withTooltip(t -> Common.newTooltip("menu.seasonhud.season.showDay.tooltip"))
         .withValues(ShowDay.getValues())
@@ -278,13 +272,24 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
                 (b, val) -> this.seasonColor = val);
     widgets.addAll(Arrays.asList(showDayButton, seasonColorButton));
 
-    row += 1; // Row 4 (enableMinimapIntegration -> Row 2)
+    row += 1;
+
     CycleButton<Boolean> showSubSeasonButton = CycleButton.onOffBuilder(showSubSeason)
         .withTooltip(t -> Common.newTooltip("menu.seasonhud.season.showSubSeason.tooltip"))
         .create(leftButtonX, (buttonStartY + (row * offsetY)), buttonWidth, buttonHeight,
                 Common.translatedText("menu.seasonhud.season.showSubSeason.button"),
                 (b, val) -> this.showSubSeason = val);
 
+    if (Common.fabricSeasonsLoaded() && this.minecraft != null) {
+      int seasonLength = Services.SEASON.currentFabricSeasonLength(this.minecraft.player);
+
+      if ((seasonLength % 3) != 0) {
+        showSubSeasonButton.active = false;
+        showSubSeasonButton.getTooltip()
+            .set(0, Common.newTooltip("menu.seasonhud.season.showSubSeason.tooltip.error", seasonLength,
+                                      seasonLength * 24000).get(0));
+      }
+    }
     widgets.add(showSubSeasonButton);
 
     if (Common.hasTropicalSeasons()) {
@@ -296,9 +301,12 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
 
       widgets.add(showTropicalSeasonButton);
     }
+  }
 
+  private void calendarButtons() {
     if (Common.hasCalendarLoaded()) {
-      row += 1; // Row 5 ((enableMinimapIntegration -> Row 3)
+      row += 1;
+
       CycleButton<Boolean> needCalendarButton = CycleButton.onOffBuilder(needCalendar)
           .withTooltip(t -> {
             if (!Common.clientSideConfig(this.minecraft)) {
@@ -333,9 +341,12 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
 
       widgets.addAll(Arrays.asList(needCalendarButton, calendarDetailModeButton));
     }
+  }
 
+  private void fertilityButtons() {
     if (Common.sereneSeasonsLoaded()) {
       row += 1;
+
       CycleButton<Boolean> showFertilityButton = CycleButton.onOffBuilder(showFertility)
           .withTooltip(t -> Common.newTooltip("menu.seasonhud.season.showFertility.tooltip"))
           .create(leftButtonX, (buttonStartY + (row * offsetY)), buttonWidth, buttonHeight,
@@ -350,9 +361,13 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
 
       widgets.addAll(Arrays.asList(showFertilityButton, fertilityReplacesSeasonButton));
     }
+  }
 
+  private void fabricSeasonsButtons() {
     if (Common.fabricSeasonsLoaded()) {
-      row += 2; // Row 4 (enableMinimapIntegration -> Row 2)
+      row += 2;
+      fabricSeasonsRow = row;
+
       dayLengthBox = new EditBox(this.font, leftButtonX + 1, (buttonStartY + (row * offsetY)), buttonWidth - 2,
                                  buttonHeight, Common.literalText(String.valueOf(dayLength)));
       dayLengthBox.setMaxLength(10);
@@ -378,6 +393,19 @@ public class DisplayOptionsScreen extends SeasonHudScreen {
 
       widgets.add(dayLengthBox);
     }
+  }
+
+  @Override
+  public void init() {
+    loadConfig();
+    super.init();
+
+    row = -1;
+    defaultHudButtons();
+    seasonButtons();
+    calendarButtons();
+    fertilityButtons();
+    fabricSeasonsButtons();
 
     widgets.forEach(this::addButton);
   }