+++ common/src/main/java/club/iananderson/seasonhud/config/SeasonHudClient.java
@@ -34,8 +34,6 @@ public class SeasonHudClient {
   private static ConfigValue<Boolean> fertilityReplacesSeason;
   private static ConfigValue<Boolean> enableMinimapIntegration;
   private static ConfigValue<Boolean> showDefaultWhenMinimapHidden;
-  private static ConfigValue<Boolean> journeyMapAboveMap;
-  private static ConfigValue<Boolean> journeyMapMacOs;
 
   static {
     ForgeConfigSpec.Builder builder = new ForgeConfigSpec.Builder();
@@ -192,23 +190,6 @@ public class SeasonHudClient {
                                                                      + Client.DEFAULT_SHOW_DEFAULT_WHEN_MINIMAP_HIDDEN
                                                                      + "."))
         .define("enable_show_minimap_hidden", Client.DEFAULT_SHOW_DEFAULT_WHEN_MINIMAP_HIDDEN);
-
-    builder.push("Journeymap");
-    journeyMapAboveMap = builder.comment(StringLine.builder()
-                                             .addLine(
-                                                 "Display the season above the JourneyMap minimap, instead of below")
-                                             .addLine("(true/false)")
-                                             .lastLine("Default is " + Client.DEFAULT_JOURNEYMAP_ABOVE_MAP + "."))
-        .define("enable_above_map", Client.DEFAULT_JOURNEYMAP_ABOVE_MAP);
-
-    journeyMapMacOs = builder.comment(StringLine.builder()
-                                          .addLine("Toggle for macOS retina display scaling when using JourneyMap.")
-                                          .addLine("Enable this us using a retina display and the season line is")
-                                          .addLine("rendering around the halfway point of the screen.")
-                                          .addLine("(true/false)")
-                                          .lastLine("Default is " + Client.DEFAULT_JOURNEYMAP_MAC_OS + "."))
-        .define("enable_macOS", Client.DEFAULT_JOURNEYMAP_MAC_OS);
-    builder.pop();
     builder.pop();
     builder.pop();
   }
@@ -232,7 +213,6 @@ public class SeasonHudClient {
     configOptions.put(showDay.getPath(), Client.DEFAULT_SHOW_DAY);
     configOptions.put(enableMinimapIntegration.getPath(), Client.DEFAULT_ENABLE_MINIMAP_INTEGRATION);
     configOptions.put(showDefaultWhenMinimapHidden.getPath(), Client.DEFAULT_SHOW_DEFAULT_WHEN_MINIMAP_HIDDEN);
-    configOptions.put(journeyMapAboveMap.getPath(), Client.DEFAULT_JOURNEYMAP_ABOVE_MAP);
 
     return configOptions.get(config.getPath());
   }
@@ -401,22 +381,5 @@ public class SeasonHudClient {
 
   public static void setEnableMinimapIntegration(boolean enable) {
     SeasonHudClient.enableMinimapIntegration.set(enable);
-  }
-
-  // Journeymap
-  public static boolean getJourneyMapAboveMap() {
-    return (Boolean) getOrDefault(journeyMapAboveMap);
-  }
-
-  public static void setJourneyMapAboveMap(boolean enable) {
-    SeasonHudClient.journeyMapAboveMap.set(enable);
-  }
-
-  public static boolean getJourneyMapMacOs() {
-    return (Boolean) getOrDefault(journeyMapMacOs);
-  }
-
-  public static void setJourneyMapMacOs(boolean enable) {
-    SeasonHudClient.journeyMapMacOs.set(enable);
   }
 }