+++ forge/src/main/java/club/iananderson/seasonhud/forge/client/overlays/JourneyMap.java
@@ -1,17 +1,12 @@
 package club.iananderson.seasonhud.forge.client.overlays;
 
-import club.iananderson.seasonhud.impl.minimap.CurrentMinimap;
-import club.iananderson.seasonhud.impl.minimap.mods.MinimapMods;
+import club.iananderson.seasonhud.client.overlays.JourneyMapCommon;
 import com.mojang.blaze3d.vertex.PoseStack;
-import journeymap.client.render.draw.DrawUtil;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiComponent;
 
 public class JourneyMap extends GuiComponent {
   public static JourneyMap HUD_INSTANCE;
-
-  public JourneyMap() {
-  }
 
   public static void init() {
     HUD_INSTANCE = new JourneyMap();
@@ -20,14 +15,6 @@ public class JourneyMap extends GuiComponent {
   public void render(PoseStack graphics) {
     Minecraft mc = Minecraft.getInstance();
 
-    if (CurrentMinimap.journeyMapLoaded() && CurrentMinimap.shouldDrawMinimapHud(MinimapMods.JOURNEYMAP, mc)) {
-      JourneyMapCommon journeyMapCommon = JourneyMapCommon.getInstance(Minecraft.getInstance());
-
-      graphics.pushPose();
-      graphics.scale(1 / journeyMapCommon.getFontScale(), 1 / journeyMapCommon.getFontScale(), 0);
-      DrawUtil.sizeDisplay(journeyMapCommon.getScreenWidth(), journeyMapCommon.getScreenHeight());
-      graphics.popPose();
-      journeyMapCommon.drawSeasonLabel(graphics);
-    }
+    JourneyMapCommon.renderHud(graphics, mc);
   }
 }