+++ forge/src/main/java/club/iananderson/seasonhud/forge/event/ClientEvents.java
@@ -1,8 +1,8 @@
 package club.iananderson.seasonhud.forge.event;
 
 import club.iananderson.seasonhud.Common;
-import club.iananderson.seasonhud.SeasonHudClientCommon;
 import club.iananderson.seasonhud.client.KeyBindings;
+import club.iananderson.seasonhud.client.SeasonHudClientCommon;
 import club.iananderson.seasonhud.forge.client.overlays.JourneyMap;
 import club.iananderson.seasonhud.forge.client.overlays.MapAtlases;
 import club.iananderson.seasonhud.forge.client.overlays.SeasonHudOverlay;