+++ forge/src/main/java/club/iananderson/seasonhud/forge/platform/ForgeSeasonHelper.java
@@ -207,6 +207,17 @@ public class ForgeSeasonHelper implements SeasonHelper {
   }
 
   @Override
+  public Component eclipticSeasonComponent(Player player) {
+    ResourceKey<Level> currentDim = Objects.requireNonNull(Minecraft.getInstance().level).dimension();
+
+    if (validEclipticSeasonsDim(currentDim)) {
+      return EclipticUtil.INSTANCE.getSolarTerm(player.level).getTranslation();
+    } else {
+      return Seasons.NULL.getSeasonNameTranslated();
+    }
+  }
+
+  @Override
   public SubSeasons currentEclipticSubSeason(Player player) {
     ResourceKey<Level> currentDim = Objects.requireNonNull(Minecraft.getInstance().level).dimension();
 
@@ -256,7 +267,7 @@ public class ForgeSeasonHelper implements SeasonHelper {
     int duration = CommonConfig.Season.lastingDaysOfEachTerm.get() * 6;
 
     if (SeasonHudClient.getShowSubSeason() && Calendar.validDetailedMode(player)) {
-      duration /= 3; // 3 sub-season per season
+      duration /= 6; // 6 sub-season per season
     }
 
     return duration;